requirements.txt
Flask
Flask-Cors
google-auth
google-auth-oauthlib
google-cloud
google-cloud-aiplatform
langchain
requests
langchain_google_vertexai

#new
Flask
flask-cors
PyMuPDF
Werkzeug
langchain-google-vertexai
google-auth

